Refactor and rename component/title
