+++ lib/galaxy/webapps/galaxy/buildapp.py
@@ -277,7 +277,7 @@ def app_pair(global_conf, load_app_kwds=None, wsgi_preflight=True, **kwargs):
     webapp.add_client_route("/interactivetool_entry_points/list")
     webapp.add_client_route("/libraries{path:.*?}")
     webapp.add_client_route("/storage{path:.*?}")
-    webapp.add_client_route("/genome-list")
+    webapp.add_client_route("/dbkeys")
 
     # ==== Done
     # Indicate that all configuration settings have been provided
